Updated bongo cat RLE frame generation.
This now encodes both sequences of duplicate values and also sequences
of unique values into our RLE format, for improved data packing.

As before, the Makefile generates the RLE data if you change the
original raw data.  Arrays in PROGMEM are now shrunk as far as possible,
and have been moved into #includable header files (I guess I should
probably do that for the decoding function too, so it's easier for other
folks to make use of them!)

In this, the bongo cat animation is played on the *slave* side, and is
animated based upon synchronised WPM from QMK's WPM feature.  This part
only really works well with my rewritten WPM system in QMK's develop
branch (but I'm told that it's going to make its way to master soon,
yay!), but it could be tied to other triggers just as easily.